src/test/java/ru/druzhininyy/leetcode/exercises/algorithms/problem0017/SolutionTest.java
package ru.druzhininyy.leetcode.exercises.algorithms.problem0017;

import com.fasterxml.jackson.core.type.TypeReference;
import com.fasterxml.jackson.databind.ObjectMapper;
import lombok.Getter;
import lombok.Setter;
import lombok.ToString;
import org.junit.jupiter.api.BeforeAll;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.io.IOException;
import java.util.Comparator;
import java.util.List;
import java.util.stream.Stream;

import static org.junit.jupiter.api.Assertions.*;

@DisplayName("Test: 17. Letter Combinations of a Phone Number.")
public class SolutionTest {

    @Getter
    @Setter
    @ToString
    public static class TestCase {
        public String digits;
        public List<String> expected;
    }

    private static final String SOURCE_FILE = "problem0017.json";
    public static List<TestCase> testCases;

    @BeforeAll
    public static void loadTestCases() {
        ObjectMapper objectMapper = new ObjectMapper();
        try {
            testCases = objectMapper.readValue(
                    SolutionTest.class.getClassLoader().getResourceAsStream(SOURCE_FILE),
                    new TypeReference<>() {
                    });
        } catch (IOException e) {
            System.out.println("Error while loading data from " + SOURCE_FILE);
            System.out.println(e.getMessage());
        }
    }

    public static Stream<Arguments> provideTestsArguments() {
        return testCases.stream().map(Arguments::of);
    }

    @ParameterizedTest
    @MethodSource("provideTestsArguments")
    public void runTestCases(TestCase testCase) {
        System.out.println(testCase);
        testCase.getExpected().sort(Comparator.naturalOrder());
        var actual = Solution.letterCombinations(testCase.digits);
        assertEquals(testCase.expected, actual);
    }

}



